é parece que terminou

+++ ucdb/view/struct/components/Login/login_features.js
@@ -1,3 +1,8 @@
+/**
+ * @Author: Rafael da Silva Pereira Matricula: 117110921. UFCG: Ciência da Computação.
+ * @Author: Áthila Matheus Barros Borges Matricula: 118210206. UFCG: Ciência da Computação.
+ * Modulo que controla o login do usuario.
+ */
 import {postData} from "../../../../controller/rest_controller.js";
 import {loginButton} from "../../../scripts/renderButtons.js";
 export {loginFeatures}
@@ -16,22 +21,38 @@ function loginFeatures(local,applicationHeader) {
             email: applicationHeader._form.email.value,
             password: applicationHeader._form.psw.value
         };
-        postData("login/", requestUser)
-            .then(response => {
-                const tokenValue = response.token;
-                window.localStorage.setItem("___access_token___", tokenValue);
-                window.localStorage.setItem("___user_email___", requestUser.email);
-                setTimeout(() => {
+        let valid = true;
+        if (requestUser.email === "") {
+            valid = false;
+            alert("Email não valido");
+        } else if (!requestUser.email.includes("@")) {
+            valid = false;
+            alert("Email não valido");
+        } else if (requestUser.password === "") {
+            valid = false;
+            alert("Sua senha deve pelo menos 8 caracteres");
+        } else if (requestUser.password.length < 8) {
+            valid = false;
+            alert("Sua senha deve pelo menos 8 caracteres");
+        }
+        if (valid) {
+            postData("login/", requestUser)
+                .then(response => {
+                    const tokenValue = response.token;
+                    window.localStorage.setItem("___access_token___", tokenValue);
+                    window.localStorage.setItem("___user_email___", requestUser.email);
+                    setTimeout(() => {
                         alert("seu login ira expirar em menos de 5 minutos, " +
                             "salve seu progresso que em menos de 5 minutos\nestaremos refazendo o seu login");
                         window.localStorage.clear();
-                }, 1000 * 60 * 55);
+                    }, 1000 * 60 * 55);
 
-                setTimeout(()=> {
-                    loginButton(local, applicationHeader)
-                }, 1000 * 60 * 60); // recriando o token
+                    setTimeout(() => {
+                        loginButton(local, applicationHeader)
+                    }, 1000 * 60 * 60); // recriando o token
 
-                applicationHeader._closeButton.click();
-            }).catch(err => alert("algo de errado aconteceu, por favor tente novamente mais tarde"));
-    };
+                    applicationHeader._closeButton.click();
+                }).catch(() => alert("Não foi possivel fazer o login, por favor tarde"));
+        }
+    }
 }
